Rationalize code
